Fix #2389: initial configuration instance is not deleted twice
This situation occurs when 2 request in more/less the same time
were sent into Atmo. Than in first request configuration instance is
deleted and second one has nil instead of configuration.

After adding check if configuration instance exists problem seems to be
solved.

+++ app/models/appliance.rb
@@ -67,8 +67,7 @@ class Appliance < ActiveRecord::Base
   private
 
   def remove_appliance_configuration_instance_if_needed
-    if appliance_configuration_instance.appliances.blank?
-      Air.action_logger.info "         -- appliance configuration intance #{appliance_configuration_instance.id} destroyed by appliance #{id}"
+    if appliance_configuration_instance && appliance_configuration_instance.appliances.blank?
       appliance_configuration_instance.destroy
     end
   end
